kbuild: allow setting zstd compression level for kernel
In a similar way to the first patch in this series, add another Kconfig
option that allows setting the zstd compression level for the kernel
as well.
This enables users to choose either a compression level between 1 and 19,
or instead use the highest possible level (ultra-22).
Also add '-T0' parameter in both cases.

Signed-off-by: torvic9 <torvic9@mailbox.org>
Signed-off-by: Dark-Matter7232 <kerneldeveloper7232@gmail.com>