maple_tree: Reorganize testing to restore module testing
Along the development cycle, the testing code support for
module/in-kernel compiles was removed.  Restore this functionality by
moving any internal API tests to the userspace side, as well as
threading tests.  Fix the lockdep issues and add a way to reduce memory
usage so the tests can complete with KASAN + memleak detection.

Signed-off-by: Liam R. Howlett <Liam.Howlett@oracle.com>